+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Fix phx-remove on sticky LiveViews
   * Fix phx-disabled-with not restoring button disabled state on empty diff acknowledgement
   * Fix stream reset not ordering items correctly
+  * Send `{:shutdown, :cancel}` to `handle_async/3` on `cancel_async`
 
 ### Enhancements
   * Add `JS.toggle_class`